requirements.txt
llama-index==0.10.68
llama-index-vector-stores-chroma==0.1.10
llama-index-llms-groq==0.1.4
llama-index-embeddings-huggingface==0.2.3
chromadb==0.5.15
groq==0.9.0
sentence-transformers==3.0.1
biopython==1.84
python-dotenv==1.0.1
tenacity==8.3.0
requests==2.32.3
pytest==8.3.2
pytest-mock==3.14.0
